+++ src/ui/src/lib/NodesTable.svelte
@@ -32,7 +32,7 @@
 	}
 </script>
 
-<div class="table-container p-2">
+<div class="p-2">
 	{#if $nodes}
 		<VersionPicker bind:updateMethod={$updateMethod} bind:flavor={$flavor} bind:version={$version} bind:artifact={$artifact} />
 		<SvelteTable {columns} rows={$nodes} classNameTable="table table-hover table-compact" on:clickRow={onRowClick} sortBy="id" />
